+++ Tests/CIOV2ClientTests.m
@@ -7,6 +7,7 @@
 //
 
 #import "CIOV2Client.h"
+#import "TestUtil.h"
 #import <XCTest/XCTest.h>
 
 @interface CIOV2ClientTests : XCTestCase
@@ -27,16 +28,6 @@
     [super tearDown];
     [self.client clearCredentials];
 }
-
-#define AssertRequestPathTypeMethod(request, requestPath, type, requestMethod) \
-    do { \
-        CIORequest *_req = (request); \
-        if (![_req isKindOfClass:[type class]]) { \
-            XCTFail(@"%@ is wrong type", _req); \
-        } \
-        XCTAssertEqualObjects(_req.path, (requestPath)); \
-        XCTAssertEqualObjects(_req.method, (requestMethod)); \
-    } while (0)
 
 #pragma mark Account