+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CIOAPIClient CHANGELOG
 
+## 1.0
+
+* All request parameters are now encapsulated in dedicated request objects which define exactly which parameters are allowed for each request. The documentation for these objects mirrors the Context.IO API documentation.
+* Lite API Support
+  - This required breaking the previous `CIOAPISession` class in to `CIOV2Client` and `CIOLiteClient` classes, `CIOAPIClient` is now an abstract base class which handles shared logic of the two APIs.
+  - `CIOAPIClient` now keeps an internal `CIOAPISession` which is used for executing requests, rather than having `CIOAPISession` subclass `CIOAPIClient`.
+* Carthage support
+* Massive increase in documentation
+* Playground is now part of the main `CIOAPIClient.xcworkspace`
+* Extensive test coverage
+
 ## 0.9.1
 
 * Documentation improvements